backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.6
django-allauth==0.45.0
django-cors-headers==3.8.0
django-rest-framework==0.1.0
django-sslserver==0.22
djangorestframework==3.12.4
ecdsa==0.17.0
idna==3.2
jose==1.0.0
oauthlib==3.1.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
